requirements.txt
asgiref==3.5.2
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
cloudinary==1.29.0
cryptography==38.0.1
Django==4.1.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
idna==3.4
Pillow==9.2.0
pycparser==2.21
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
